refactor: 优化代码结构和变量声明以提高可读性和一致性
- 调整部分类的构造函数和静态方法的访问修饰符
- 重构变量声明位置以避免重复导入
- 修改部分逻辑以使用更简洁的代码实现
- 优化 Redis 锁释放脚本的返回类型声明
- 将内部类声明为 static 以支持独立实例化
- 统一代码格式，调整缩进和换行以提高可读性

+++ rose-spring-boot/rose-spring-boot-redis/src/main/java/io/github/rose/redis/service/RedisService.java
@@ -15,9 +15,6 @@
  */
 package io.github.rose.redis.service;
 
-import java.nio.charset.Charset;
-import java.util.*;
-import java.util.concurrent.TimeUnit;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.connection.RedisConnection;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
@@ -27,6 +24,10 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.ScanOptions;
 import org.springframework.data.redis.core.script.DefaultRedisScript;
 import org.springframework.data.redis.core.script.RedisScript;
+
+import java.nio.charset.Charset;
+import java.util.*;
+import java.util.concurrent.TimeUnit;
 
 /**
  * @author zhijun.chen
@@ -129,9 +130,9 @@ public class RedisService {
 
     public boolean releaseLock(String lockKey, String value) {
         String script = "if mq.call('get', KEYS[1]) == ARGV[1] then return mq.call('del', KEYS[1]) else return 0 end";
-        RedisScript<Long> redisScript = new DefaultRedisScript<>(script, Long.class);
+        RedisScript<Integer> redisScript = new DefaultRedisScript<>(script, Integer.class);
         return redisTemplate
-                .execute(redisScript, Collections.singletonList(lockKey), value)
-                .equals(1);
+            .execute(redisScript, Collections.singletonList(lockKey), value)
+            .equals(1);
     }
 }
